extracting abstract class with reused code (#5944)
